# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Restaurant:
    properties:
      category:
        type: string
      full_address:
        type: string
      id:
        type: integer
      lat:
        description: 'TODO: add validation'
        type: string
      lng:
        description: 'TODO: add validation'
        type: string
      name:
        type: string
      position:
        type: integer
      price_range:
        type: string
      ratings:
        type: integer
      score:
        type: number
      zip_code:
        description: 'TODO: Add additional validation for the zip range'
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: valentyna.koshelnyk@deliveryhero.com
    name: PandaEats API
  description: A simple food ordering app
  title: PandaEats API
  version: "1.0"
paths:
...
  /restaurants:
    get:
      description: Retrieves the list of all restaurants from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Restaurant'
            type: array
      summary: Get all restaurants
    post:
      consumes:
      - application/json
      description: Adds a new restaurant to the restaurants table
      produces:
      - application/json
      responses:
        "201":
          description: Restaurant created
          schema:
            type: string
        "400":
          description: Error reading request body or decoding restaurant
          schema:
            type: string
        "422":
          description: Error creating restaurant
          schema:
            type: string
      summary: Adds a new restaurant
  /restaurants/{id}:
    delete:
      description: Deletes the restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Deletes a restaurant record
    put:
      consumes:
      - application/json
      description: Updates info about the restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      - description: Restaurant
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/entity.Restaurant'
      produces:
      - application/json
      responses:
        "204":
          description: Successfully updated restaurant
          schema:
            type: string
        "400":
          description: Error updating restaurant
          schema:
            type: string
        "409":
          description: Error decoding restaurant
          schema:
            type: string
      summary: Updates a restaurant information
...
swagger: "2.0"